+++ app/Http/Controllers/Api/V1/ColorsController.php
@@ -13,6 +13,7 @@ use App\Jobs\UpdateColor;
 use App\Section;
 use App\SectionType;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
+use Illuminate\Support\Facades\Auth;
 
 class ColorsController extends Controller
 {
@@ -20,6 +21,11 @@ class ColorsController extends Controller
     public function createColor(CreateColorRequest $request, $bridgeId)
     {
         try{
+            $user = Auth::user();
+            $bridge = Bridge::findOrFail($bridgeId);
+            if($user->id !== $bridge->user_id)
+                throw new ModelNotFoundException();
+
             $sectionType = SectionType::where('name', SectionType::COLORS)->get()->first();
             $section = Section::where('section_type_id', $sectionType->id)->where('bridge_id', $bridgeId)->get()->first();
 
@@ -47,6 +53,11 @@ class ColorsController extends Controller
     public function updateColor(CreateColorRequest $request, $bridgeId, $colorId)
     {
         try{
+            $user = Auth::user();
+            $bridge = Bridge::findOrFail($bridgeId);
+            if($user->id !== $bridge->user_id)
+                throw new ModelNotFoundException();
+
             (new UpdateColor($request->only(['hex', 'cmyk', 'rgb']), $bridgeId, Color::findOrFail($colorId)))->handle();
 
             $bridge = Bridge::with('sections', 'icons', 'icons.converted', 'images', 'images.converted', 'fonts', 'fonts.variant', 'fonts.variant.fontFamily', 'colors')->findOrFail($bridgeId);
@@ -71,10 +82,16 @@ class ColorsController extends Controller
     public function deleteColor($bridgeId, $colorId)
     {
         try{
+
+            $user = Auth::user();
+            $bridge = Bridge::findOrFail($bridgeId);
+            if($user->id !== $bridge->user_id)
+                throw new ModelNotFoundException();
+
             $color = Color::findOrFail($colorId);
             $color->delete();
 
-            $bridge = Bridge::with('sections', 'icons', 'icons.converted', 'images', 'images.converted', 'fonts', 'fonts.variant', 'fonts.variant.fontFamily', 'colors')->findOrFail($bridgeId);
+            $bridge = Bridge::with('sections', 'icons', 'icons.converted', 'images', 'images.converted', 'fonts', 'fonts.variant', 'fonts.variant.fontFamily', 'colors')->where('user_id', $user->id)->findOrFail($bridgeId);
             try{
                 event(new BridgeUpdated($bridge));
             }catch(\Exception $e){}
